requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2.3
django-extensions==4.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
PyJWT==2.9.0
python-decouple==3.8
requests==2.32.4
sqlparse==0.5.3
stripe==12.2.0
typing_extensions==4.14.0
urllib3==2.5.0
